+++ command/cp.go
@@ -181,6 +181,10 @@ func NewSharedFlags() []cli.Flag {
 			Name:  "content-encoding",
 			Usage: "set content encoding for target: defines content encoding header for object, e.g. --content-encoding gzip",
 		},
+		&cli.StringFlag{
+			Name:  "content-disposition",
+			Usage: "set content disposition for target: defines content disposition header for object, e.g. --content-disposition 'attachment; filename=\"filename.jpg\"'",
+		},
 		&cli.IntFlag{
 			Name:        "no-such-upload-retry-count",
 			Usage:       "number of times that a request will be retried on NoSuchUpload error; you should not use this unless you really know what you're doing",
@@ -284,7 +288,7 @@ type Copy struct {
 	contentEncoding       string
 	showProgress          bool
 	progressbar           progressbar.ProgressBar
-
+	contentDisposition    string
 	// region settings
 	srcRegion string
 	dstRegion string
@@ -348,6 +352,7 @@ func NewCopy(c *cli.Context, deleteSource bool) (*Copy, error) {
 		showProgress:          c.Bool("show-progress"),
 		progressbar:           commandProgressBar,
 
+		contentDisposition:    c.String("content-disposition"),
 		// region settings
 		srcRegion: c.String("source-region"),
 		dstRegion: c.String("destination-region"),
@@ -653,7 +658,9 @@ func (c Copy) doUpload(ctx context.Context, srcurl *url.URL, dsturl *url.URL) er
 	if c.contentEncoding != "" {
 		metadata.SetContentEncoding(c.contentEncoding)
 	}
-
+	if c.contentDisposition != "" {
+		metadata.SetContentDisposition(c.contentDisposition)
+	}
 	reader := &CountingReader{
 		pb:      c.progressbar,
 		fp:      file,
@@ -714,6 +721,9 @@ func (c Copy) doCopy(ctx context.Context, srcurl, dsturl *url.URL) error {
 	}
 	if c.contentEncoding != "" {
 		metadata.SetContentEncoding(c.contentEncoding)
+	}
+	if c.contentDisposition != "" {
+		metadata.SetContentDisposition(c.contentDisposition)
 	}
 
 	err = c.shouldOverride(ctx, srcurl, dsturl)